requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
cfgv==3.0.0
Click==7.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-filter==2.2.0
django-stubs==1.4.0
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
identify==1.4.11
importlib-metadata==1.5.0
Markdown==3.2.1
mccabe==0.6.1
mypy==0.761
mypy-extensions==0.4.3
nodeenv==1.3.5
pathspec==0.7.0
pre-commit==2.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
virtualenv==20.0.5
zipp==3.0.0
